PR comments

+++ cmd/root.go
@@ -127,7 +127,7 @@ any client SSL certificates.`,
 		"Path to a service account key to use for authentication.")
 	cmd.PersistentFlags().BoolVarP(&c.conf.GcloudAuth, "gcloud-auth", "g", false,
 		"Use gcloud's user configuration to retrieve a token for authentication.")
-	cmd.PersistentFlags().Uint32Var(&c.conf.MaxConnections, "max-connections", 0,
+	cmd.PersistentFlags().Uint64Var(&c.conf.MaxConnections, "max-connections", 0,
 		`Limits the number of connections by refusing any additional connections.
 When this flag is not set, there is no limit.`)
 	cmd.PersistentFlags().StringVar(&c.telemetryProject, "telemetry-project", "",
